Refactor 0004 async migration: make rollbacks robust

+++ posthog/async_migrations/migrations/0004_replicated_schema.py
@@ -1,7 +1,7 @@
 import re
 from dataclasses import dataclass
 from functools import cached_property
-from typing import List, Optional, cast
+from typing import Dict, Optional, cast
 
 import structlog
 from constance import config
@@ -14,6 +14,7 @@ from posthog.async_migrations.definition import (
     AsyncMigrationOperation,
     AsyncMigrationOperationSQL,
 )
+from posthog.utils import flatten
 
 logger = structlog.get_logger(__name__)
 
@@ -30,10 +31,10 @@ The migration strategy:
     1. We have a list of tables that might need replacing below.
     2. For each one, we replace the current engine with the appropriate Replicated by:
         a. creating a new table with the right engine and identical schema
-        b. temporarily stopping ingestion to the table
+        b. temporarily stopping ingestion to the table by dropping the kafka table
         c. using `ALTER TABLE ATTACH/DROP PARTITIONS` to move data to the new table.
         d. rename tables
-        e. re-enabling ingestion
+    3. Once all tables are updated, we create needed distributed tables and re-enable ingestion
 
 We use ATTACH/DROP tables to do the table migration instead of a normal INSERT. This method allows
 moving data without increasing disk usage between identical schemas.
@@ -41,8 +42,9 @@ moving data without increasing disk usage between identical schemas.
 `events` and `session_recording_events` require extra steps as they're also sharded:
 
     1. The new table should be named `sharded_TABLENAME`
-    2. Create `TABLENAME` and `writable_TABLENAME` tables which are responsible for distributed reads and writes
-    3. Update materialized views to write to `writable_TABLENAME`
+    2. When re-enabling ingestion, we create `TABLENAME` and `writable_TABLENAME` tables
+       which are responsible for distributed reads and writes.
+    3. We re-create materialized views to write to `writable_TABLENAME`
 
 Constraints:
 
@@ -51,6 +53,8 @@ Constraints:
     3. This migration requires there to be no ongoing part merges while it's executing.
     4. This migration depends on 0002_events_sample_by. If it didn't, this could be a normal migration.
     5. This migration depends on the person_distinct_id2 async migration to have completed.
+    6. We can't stop ingestion by dropping/detaching materialized view as we can't restore to the right (non-replicated) schema afterwards.
+    7. Async migrations might fail _before_ a step executes and rollbacks need to account for that, which complicates renaming logic.
 """
 
 
@@ -58,8 +62,8 @@ Constraints:
 class TableMigrationData:
     name: str
     new_table_engine: MergeTreeEngine
-    materialized_view_name: Optional[str]
-    create_materialized_view: Optional[str]
+    kafka_table_name: Optional[str]
+    create_kafka_table: Optional[str]
 
     @property
     def renamed_table_name(self):
@@ -77,7 +81,9 @@ class TableMigrationData:
 @dataclass(frozen=True)
 class ShardedTableMigrationData(TableMigrationData):
     rename_to: str
-    extra_tables: List[str]
+    extra_tables: Dict[str, str]
+    materialized_view_name: str
+    create_materialized_view: str
 
     @property
     def renamed_table_name(self):
@@ -108,9 +114,12 @@ class Migration(AsyncMigrationDefinition):
 
     @cached_property
     def operations(self):
-        TABLE_MIGRATION_OPERATIONS = [
-            operation for table in self.tables_to_migrate() for operation in self.replicated_table_operations(table)
-        ]
+        TABLE_MIGRATION_OPERATIONS = list(
+            flatten([list(self.replicated_table_operations(table)) for table in self.tables_to_migrate()])
+        )
+        RE_ENABLE_INGESTION_OPERATIONS = list(
+            flatten([list(self.finalize_table_operations(table)) for table in self.tables_to_migrate()])
+        )
 
         return [
             AsyncMigrationOperationSQL(sql="SYSTEM STOP MERGES", rollback="SYSTEM START MERGES"),
@@ -119,6 +128,7 @@ class Migration(AsyncMigrationDefinition):
                 rollback_fn=lambda _: setattr(config, "COMPUTE_MATERIALIZED_COLUMNS_ENABLED", True),
             ),
             *TABLE_MIGRATION_OPERATIONS,
+            *RE_ENABLE_INGESTION_OPERATIONS,
             AsyncMigrationOperation(
                 fn=lambda _: setattr(config, "COMPUTE_MATERIALIZED_COLUMNS_ENABLED", False),
                 rollback_fn=lambda _: setattr(config, "COMPUTE_MATERIALIZED_COLUMNS_ENABLED", True),
@@ -135,9 +145,9 @@ class Migration(AsyncMigrationDefinition):
             rollback=f"DROP TABLE IF EXISTS {table.tmp_table_name}",
         )
 
-        if table.materialized_view_name is not None:
+        if table.kafka_table_name is not None:
             yield AsyncMigrationOperationSQL(
-                sql=f"DROP TABLE IF EXISTS {table.materialized_view_name}", rollback=table.create_materialized_view,
+                sql=f"DROP TABLE IF EXISTS {table.kafka_table_name}", rollback=cast(str, table.create_kafka_table)
             )
 
         yield AsyncMigrationOperation(
@@ -147,26 +157,33 @@ class Migration(AsyncMigrationDefinition):
 
         yield AsyncMigrationOperation(
             fn=lambda _: self.rename_tables(
-                table.name, table.tmp_table_name, table.renamed_table_name, table.backup_table_name
+                [table.name, table.backup_table_name], [table.tmp_table_name, table.renamed_table_name],
             ),
             rollback_fn=lambda _: self.rename_tables(
-                table.backup_table_name,
-                table.renamed_table_name,
-                table.tmp_table_name,
-                table.name,
-                skip_unless_backup_exists=True,
+                [table.renamed_table_name, table.tmp_table_name],
+                [table.backup_table_name, table.name],
+                verify_table_exists=table.backup_table_name,
             ),
         )
 
-        if table.materialized_view_name is not None:
+    def finalize_table_operations(self, table: TableMigrationData):
+        # NOTE: Relies on IF NOT EXISTS on the query
+        if isinstance(table, ShardedTableMigrationData):
+            for table_name, create_table_query in table.extra_tables.items():
+                yield AsyncMigrationOperationSQL(sql=create_table_query, rollback=f"DROP TABLE IF EXISTS {table_name}")
+
+        if isinstance(table, ShardedTableMigrationData) and table.materialized_view_name is not None:
+            yield AsyncMigrationOperationSQL(sql=f"DROP TABLE IF EXISTS {table.materialized_view_name}",)
+
+        if table.kafka_table_name is not None:
+            yield AsyncMigrationOperationSQL(
+                sql=cast(str, table.create_kafka_table), rollback=f"DROP TABLE IF EXISTS {table.kafka_table_name}"
+            )
+
+        if isinstance(table, ShardedTableMigrationData) and table.materialized_view_name is not None:
             yield AsyncMigrationOperationSQL(
                 sql=table.create_materialized_view, rollback=f"DROP TABLE IF EXISTS {table.materialized_view_name}",
             )
-
-        # NOTE: Relies on IF NOT EXISTS on the query
-        if isinstance(table, ShardedTableMigrationData):
-            for create_table_query in table.extra_tables:
-                yield AsyncMigrationOperationSQL(sql=create_table_query)
 
     def get_current_engine(self, table_name: str) -> Optional[str]:
         result = sync_execute(
@@ -229,24 +246,16 @@ class Migration(AsyncMigrationDefinition):
             sync_execute(f"ALTER TABLE {to_table} REPLACE PARTITION {partition} FROM {from_table}")
             sync_execute(f"ALTER TABLE {from_table} DROP PARTITION {partition}")
 
-    def rename_tables(
-        self, data_table_name, tmp_table_name, new_main_table_name, backup_table_name, skip_unless_backup_exists=False
-    ):
+    def rename_tables(self, rename_1, rename_2, verify_table_exists=None):
         # :KLUDGE: Due to how async migrations rollback works, we need to check whether backup table exists even if the rename failed
         #   in the first place
-        if skip_unless_backup_exists and self.get_current_engine(backup_table_name) is None:
+        if verify_table_exists and self.get_current_engine(verify_table_exists) is None:
             logger.info(
-                "Backup table doesn't exist, skipping renaming.",
-                data_table_name=data_table_name,
-                new_table_name=tmp_table_name,
-                new_main_table_name=new_main_table_name,
-                backup_table_name=backup_table_name,
+                "(Rollback) Source table doesn't exist, skipping renaming.", rename_1=rename_1, rename_2=rename_2
             )
             return
 
-        return sync_execute(
-            f"RENAME TABLE {data_table_name} TO {backup_table_name}, {tmp_table_name} TO {new_main_table_name}"
-        )
+        return sync_execute(f"RENAME TABLE {rename_1[0]} TO {rename_1[1]}, {rename_2[0]} TO {rename_2[1]}")
 
     def get_number_of_nodes_in_cluster(self):
         return sync_execute(
@@ -255,27 +264,32 @@ class Migration(AsyncMigrationDefinition):
 
     def tables_to_migrate(self):
         from ee.clickhouse.sql.cohort import COHORTPEOPLE_TABLE_ENGINE
-        from ee.clickhouse.sql.dead_letter_queue import DEAD_LETTER_QUEUE_TABLE_ENGINE, DEAD_LETTER_QUEUE_TABLE_MV_SQL
+        from ee.clickhouse.sql.dead_letter_queue import (
+            DEAD_LETTER_QUEUE_TABLE_ENGINE,
+            KAFKA_DEAD_LETTER_QUEUE_TABLE_SQL,
+        )
         from ee.clickhouse.sql.events import (
             DISTRIBUTED_EVENTS_TABLE_SQL,
             EVENTS_DATA_TABLE_ENGINE,
             EVENTS_TABLE_MV_SQL,
+            KAFKA_EVENTS_TABLE_SQL,
             WRITABLE_EVENTS_TABLE_SQL,
         )
-        from ee.clickhouse.sql.groups import GROUPS_TABLE_ENGINE, GROUPS_TABLE_MV_SQL
+        from ee.clickhouse.sql.groups import GROUPS_TABLE_ENGINE, KAFKA_GROUPS_TABLE_SQL
         from ee.clickhouse.sql.person import (
-            PERSON_DISTINCT_ID2_MV_SQL,
+            KAFKA_PERSON_DISTINCT_ID2_TABLE_SQL,
+            KAFKA_PERSONS_TABLE_SQL,
             PERSON_DISTINCT_ID2_TABLE_ENGINE,
             PERSON_STATIC_COHORT_TABLE_ENGINE,
             PERSONS_TABLE_ENGINE,
-            PERSONS_TABLE_MV_SQL,
         )
         from ee.clickhouse.sql.plugin_log_entries import (
+            KAFKA_PLUGIN_LOG_ENTRIES_TABLE_SQL,
             PLUGIN_LOG_ENTRIES_TABLE_ENGINE,
-            PLUGIN_LOG_ENTRIES_TABLE_MV_SQL,
         )
         from ee.clickhouse.sql.session_recording_events import (
             DISTRIBUTED_SESSION_RECORDING_EVENTS_TABLE_SQL,
+            KAFKA_SESSION_RECORDING_EVENTS_TABLE_SQL,
             SESSION_RECORDING_EVENTS_DATA_TABLE_ENGINE,
             SESSION_RECORDING_EVENTS_TABLE_MV_SQL,
             WRITABLE_SESSION_RECORDING_EVENTS_TABLE_SQL,
@@ -287,60 +301,64 @@ class Migration(AsyncMigrationDefinition):
                 new_table_engine=EVENTS_DATA_TABLE_ENGINE(),
                 materialized_view_name="events_mv",
                 rename_to="sharded_events",
+                kafka_table_name="kafka_events",
+                create_kafka_table=KAFKA_EVENTS_TABLE_SQL(),
                 create_materialized_view=EVENTS_TABLE_MV_SQL(),
-                extra_tables=[WRITABLE_EVENTS_TABLE_SQL(), DISTRIBUTED_EVENTS_TABLE_SQL()],
+                extra_tables={"writable_events": WRITABLE_EVENTS_TABLE_SQL(), "events": DISTRIBUTED_EVENTS_TABLE_SQL()},
             ),
             ShardedTableMigrationData(
                 name="session_recording_events",
                 new_table_engine=SESSION_RECORDING_EVENTS_DATA_TABLE_ENGINE(),
+                kafka_table_name="kafka_session_recording_events",
+                create_kafka_table=KAFKA_SESSION_RECORDING_EVENTS_TABLE_SQL(),
                 materialized_view_name="session_recording_events_mv",
                 rename_to="sharded_session_recording_events",
                 create_materialized_view=SESSION_RECORDING_EVENTS_TABLE_MV_SQL(),
-                extra_tables=[
-                    WRITABLE_SESSION_RECORDING_EVENTS_TABLE_SQL(),
-                    DISTRIBUTED_SESSION_RECORDING_EVENTS_TABLE_SQL(),
-                ],
+                extra_tables={
+                    "writable_session_recording_events": WRITABLE_SESSION_RECORDING_EVENTS_TABLE_SQL(),
+                    "session_recording_events": DISTRIBUTED_SESSION_RECORDING_EVENTS_TABLE_SQL(),
+                },
             ),
             TableMigrationData(
                 name="events_dead_letter_queue",
                 new_table_engine=DEAD_LETTER_QUEUE_TABLE_ENGINE(),
-                materialized_view_name="events_dead_letter_queue_mv",
-                create_materialized_view=DEAD_LETTER_QUEUE_TABLE_MV_SQL,
+                kafka_table_name="kafka_events_dead_letter_queue",
+                create_kafka_table=KAFKA_DEAD_LETTER_QUEUE_TABLE_SQL(),
             ),
             TableMigrationData(
                 name="groups",
                 new_table_engine=GROUPS_TABLE_ENGINE(),
-                materialized_view_name="groups_mv",
-                create_materialized_view=GROUPS_TABLE_MV_SQL,
+                kafka_table_name="kafka_groups",
+                create_kafka_table=KAFKA_GROUPS_TABLE_SQL(),
             ),
             TableMigrationData(
                 name="person",
                 new_table_engine=PERSONS_TABLE_ENGINE(),
-                materialized_view_name="person_mv",
-                create_materialized_view=PERSONS_TABLE_MV_SQL,
+                kafka_table_name="kafka_person",
+                create_kafka_table=KAFKA_PERSONS_TABLE_SQL(),
             ),
             TableMigrationData(
                 name="person_distinct_id2",
                 new_table_engine=PERSON_DISTINCT_ID2_TABLE_ENGINE(),
-                materialized_view_name="person_distinct_id2_mv",
-                create_materialized_view=PERSON_DISTINCT_ID2_MV_SQL,
+                kafka_table_name="kafka_person_distinct_id2",
+                create_kafka_table=KAFKA_PERSON_DISTINCT_ID2_TABLE_SQL(),
             ),
             TableMigrationData(
                 name="plugin_log_entries",
                 new_table_engine=PLUGIN_LOG_ENTRIES_TABLE_ENGINE(),
-                materialized_view_name="plugin_log_entries_mv",
-                create_materialized_view=PLUGIN_LOG_ENTRIES_TABLE_MV_SQL,
+                kafka_table_name="kafka_plugin_log_entries",
+                create_kafka_table=KAFKA_PLUGIN_LOG_ENTRIES_TABLE_SQL(),
             ),
             TableMigrationData(
                 name="cohortpeople",
                 new_table_engine=COHORTPEOPLE_TABLE_ENGINE(),
-                materialized_view_name=None,
-                create_materialized_view=None,
+                kafka_table_name=None,
+                create_kafka_table=None,
             ),
             TableMigrationData(
                 name="person_static_cohort",
                 new_table_engine=PERSON_STATIC_COHORT_TABLE_ENGINE(),
-                materialized_view_name=None,
-                create_materialized_view=None,
+                kafka_table_name=None,
+                create_kafka_table=None,
             ),
         ]
